refactor(advanced search): move select prop. comp.
