+++ projects/dsp-ui/src/lib/search/advanced-search/resource-and-property-selection/resource-and-property-selection.component.ts
@@ -17,7 +17,7 @@ import {
     ResourcePropertyDefinition
 } from '@dasch-swiss/dsp-js';
 import { SelectResourceClassComponent } from './select-resource-class/select-resource-class.component';
-import { SelectPropertyComponent } from '../select-property/select-property.component';
+import { SelectPropertyComponent } from './select-property/select-property.component';
 import { DspApiConnectionToken } from '../../../core/core.module';
