+++ projects/dsp-ui/src/lib/search/advanced-search/resource-and-property-selection/select-property/select-property.component.spec.ts
