+++ projects/dsp-ui/src/lib/search/advanced-search/resource-and-property-selection/select-property/specify-property-value/search-boolean-value/search-boolean-value.component.html
