+++ projects/dsp-ui/src/lib/search/advanced-search/resource-and-property-selection/select-property/specify-property-value/search-decimal-value/search-decimal-value.component.spec.ts
