+++ projects/dsp-ui/src/lib/search/advanced-search/resource-and-property-selection/select-property/specify-property-value/search-link-value/search-link-value.component.spec.ts
@@ -14,7 +14,7 @@ import {
     SearchEndpointV2
 } from '@dasch-swiss/dsp-js';
 import { of } from 'rxjs';
-import { DspApiConnectionToken } from '../../../../../core/core.module';
+import { DspApiConnectionToken } from '../../../../../../core/core.module';
 import { IRI } from '../operator';
 import { SearchLinkValueComponent } from './search-link-value.component';
