+++ projects/dsp-ui/src/lib/search/index.ts
@@ -3,6 +3,6 @@ export * from './search-panel/search-panel.component';
 export * from './fulltext-search/fulltext-search.component';
 export * from './expert-search/expert-search.component';
 export * from './advanced-search/advanced-search.component';
-export * from './advanced-search/select-property/specify-property-value/operator';
+export * from './advanced-search/resource-and-property-selection/select-property/specify-property-value/operator';
 export * from './services/advanced-search-params.service';
 export * from './services/gravsearch-generation.service';
